Bot Rewrite.
A lot of the base code for the bot has been rewritten. We're experimenting now with a different approach to making bots: base command extensions. All commands now run within a separately defined class for Cogs, that way each command can be reloaded with relative ease. The API wrapper for DeepL has also been entirely re-done from scratch, however, the code rewrite is still in development so we're facing issues. Here are the main changes:

- Introduced buttons instead of reactions for the about command
- Working subscribe command that follows the support server's news channel.
- DeepL API wrapper rewrite.